\name{open.question}
\encoding{UTF-8}
\alias{open.question}
\title{Open.question (data)}
\docType{data}

\description{

Extract of the answers provided in a survey designed to better know opinions about what is most important in life.

Two open-ended questions are included in the questionnaire "What is most important to you in life?" and
"What are other very important things to you? (relaunch of the first question).
}

\usage{data(open.question)}


\format{
Data frame with 300 rows and 10 columns. The rows correspond to the respondents. The first 8 columns correspond to socio-demographic variables collected through closed questions: Gender, Age_Group, Age, Education level, Genre crossed with Age, Genre crossed with Education level, Age crossed with Education level and, finally Genre crossed with Education level and Age. Age is a quantitative variable while the other variables are qualitative. The last two columns contain the answers to the open-ended questions. 
}


\keyword{datasets}


	
