\name{Xplortext-package}
\encoding{latin1}
\alias{Xplortext-package}
\alias{Xplortext}
\docType{package}
\title{Textual Analysis}
\description{
Provides a set of functions devoted to multivariate exploratory statistics on textual data.
Classical methods such as correspondence analysis and agglomerative hierarchical clustering are available.
Chronologically constrained agglomerative hierarchical clustering enriched with labelled-by-words trees is offered.
Given a division of the corpus into parts, their characteristic words and documents are identified.
Further, accessing to 'FactoMineR' functions is very easy. Two of them are relevant in textual domain.
MFA() addresses multiple lexical table allowing applications such as dealing with multilingual corpora as well as simultaneously analyzing both open-ended and closed questions in surveys. 
CaGalt() helps to explore the relationships between lexical choices and contextual variables. See <http://www.Xplortext.org> for examples.
}
\details{
\tabular{ll}{
Package: \tab Xplortext\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2018-01-23\cr
License: GPL (>=2.0)
}
}
\author{Monica Bécue-Bertaut, Ramón Alvarez-Esteban, Josep-Anton Sánchez-
    Espigares\cr
Maintainer: \email{ramon.alvarez@unileon.es}
}
\references{
Husson F., Le S., Pages J. (2011). Exploratory Multivariate Analysis by Example Using R. Chapman & Hall/CRC. <doi:10.1080/02664763.2012.657409>.

Lebart, L., Salem, A., & Berry, L. (1998). Exploring textual data. (D. Kluwer, Ed.). <doi:10.1007/978-94-017-1525-6>.

A website \url{http://www.xplortext.org}
}
