\name{Xplortext-package}
\encoding{UTF-8}
\alias{Xplortext-package}
\docType{package}
\title{Textual Analysis}
\description{
Provides a set of functions devoted to multivariate exploratory statistics on textual data.
Classical methods such as correspondence analysis and agglomerative hierarchical clustering are available.
Chronologically constrained agglomerative hierarchical clustering enriched with labelled-by-words trees is offered.
Given a division of the corpus into parts, their characteristic words and documents are identified.
Further, accessing to 'FactoMineR' functions is very easy. Two of them are relevant in textual domain.
MFA() addresses multiple lexical table allowing applications such as dealing with multilingual corpora as well as simultaneously analyzing both open-ended and closed questions in surveys. 
See \url{https://xplortext.unileon.es} for examples.
}
\details{
\tabular{ll}{
Package: \tab Xplortext\cr
Type: \tab Package\cr
Version: \tab 1.5.1\cr
Date: \tab 2022-06-12\cr
License: GPL (>=2.0)
}
}
\author{Mónica Bécue-Bertaut, Ramón Alvarez-Esteban, Josep-Anton Sánchez-
    Espigares, Belchin Kostov\cr
Maintainer: \email{ramon.alvarez@unileon.es}
}


\references{
Bécue, M. (2019). Textual Data Science with R. Chapman & Hall/CRC. \doi{10.1201/9781315212661}.

Husson F., Lê S., Pagès J. (2011). Exploratory Multivariate Analysis by Example Using R. Chapman & Hall/CRC. \doi{10.1201/b10345}.

Lebart, L., Salem, A., & Berry, L. (1998). Exploring textual data. (D. Kluwer, Ed.).
 \doi{10.1007/978-94-017-1525-6}.

A website \url{https://xplortext.unileon.es}

}
