\name{Xplortext-package}
\encoding{latin1}
\alias{Xplortext-package}
\alias{Xplortext}
\docType{package}
\title{Textual Analysis}
\description{A set of functions devoted to multivariate exploratory statistical analysis of textual data}
\details{
\tabular{ll}{
Package: \tab Xplortext\cr
Type: \tab Package\cr
Version: \tab 1.00\cr
Date: \tab 2017-05-24\cr
License: GPL (>=2.0)
}
}
\author{Monica Bécue-Bertaut, Ramón Alvarez-Esteban, Josep-Anton Sánchez-
    Espigares\cr
Maintainer: \email{ramon.alvarez@unileon.es}
}
\references{
A website \url{http://www.xplortext.org}
}



